Remove extra React imports
