+++ src/types.ts
@@ -1,0 +1,3 @@
+export interface Params {
+  onSmartScrollStatusChange?: (isScrollEnabled: boolean) => void
+}
